FEAET: logout button on header
